LPS-58791 Move classes to their proper packages
