+++ modules/apps/bookmarks/bookmarks-service/src/com/liferay/bookmarks/exportimport/staged/model/repository/BookmarksEntryStagedModelRepository.java
@@ -12,12 +12,12 @@
  * details.
  */
 
-package com.liferay.bookmarks.exportimport.stagedmodel.repository;
+package com.liferay.bookmarks.exportimport.staged.model.repository;
 
-import com.liferay.bookmarks.model.BookmarksFolder;
-import com.liferay.bookmarks.service.BookmarksFolderLocalServiceUtil;
-import com.liferay.exportimport.stagedmodel.repository.StagedModelRepository;
-import com.liferay.exportimport.stagedmodel.repository.base.BaseStagedModelRepository;
+import com.liferay.bookmarks.model.BookmarksEntry;
+import com.liferay.bookmarks.service.BookmarksEntryLocalServiceUtil;
+import com.liferay.exportimport.staged.model.repository.StagedModelRepository;
+import com.liferay.exportimport.staged.model.repository.base.BaseStagedModelRepository;
 import com.liferay.portal.kernel.dao.orm.QueryUtil;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.trash.TrashHandler;
@@ -34,17 +34,17 @@ import org.osgi.service.component.annotations.Component;
  */
 @Component(
 	immediate = true,
-	property = {"model.class.name=com.liferay.bookmarks.model.BookmarksFolder"},
+	property = {"model.class.name=com.liferay.bookmarks.model.BookmarksEntry"},
 	service = StagedModelRepository.class
 )
-public class BookmarksFolderStagedModelRepository
-	extends BaseStagedModelRepository<BookmarksFolder> {
+public class BookmarksEntryStagedModelRepository
+	extends BaseStagedModelRepository<BookmarksEntry> {
 
 	@Override
-	public void deleteStagedModel(BookmarksFolder bookmarksFolder)
+	public void deleteStagedModel(BookmarksEntry bookmarksEntry)
 		throws PortalException {
 
-		BookmarksFolderLocalServiceUtil.deleteFolder(bookmarksFolder);
+		BookmarksEntryLocalServiceUtil.deleteEntry(bookmarksEntry);
 	}
 
 	@Override
@@ -52,54 +52,59 @@ public class BookmarksFolderStagedModelRepository
 			String uuid, long groupId, String className, String extraData)
 		throws PortalException {
 
-		BookmarksFolder bookmarksFolder = fetchStagedModelByUuidAndGroupId(
+		BookmarksEntry bookmarksEntry = fetchStagedModelByUuidAndGroupId(
 			uuid, groupId);
 
-		if (bookmarksFolder != null) {
-			deleteStagedModel(bookmarksFolder);
+		if (bookmarksEntry != null) {
+			deleteStagedModel(bookmarksEntry);
 		}
 	}
 
 	@Override
-	public BookmarksFolder fetchStagedModelByUuidAndGroupId(
+	public BookmarksEntry fetchStagedModelByUuidAndGroupId(
 		String uuid, long groupId) {
 
-		return BookmarksFolderLocalServiceUtil.
-			fetchBookmarksFolderByUuidAndGroupId(uuid, groupId);
+		return BookmarksEntryLocalServiceUtil.
+			fetchBookmarksEntryByUuidAndGroupId(uuid, groupId);
 	}
 
 	@Override
-	public List<BookmarksFolder> fetchStagedModelsByUuidAndCompanyId(
+	public List<BookmarksEntry> fetchStagedModelsByUuidAndCompanyId(
 		String uuid, long companyId) {
 
-		return BookmarksFolderLocalServiceUtil.
-			getBookmarksFoldersByUuidAndCompanyId(
+		return BookmarksEntryLocalServiceUtil.
+			getBookmarksEntriesByUuidAndCompanyId(
 				uuid, companyId, QueryUtil.ALL_POS, QueryUtil.ALL_POS,
-				new StagedModelModifiedDateComparator<BookmarksFolder>());
+				new StagedModelModifiedDateComparator<BookmarksEntry>());
 	}
 
 	@Override
 	public void restoreStagedModel(
 			PortletDataContext portletDataContext,
-			BookmarksFolder bookmarksFolder)
+			BookmarksEntry bookmarksEntry)
 		throws PortletDataException {
 
 		long userId = portletDataContext.getUserId(
-			bookmarksFolder.getUserUuid());
+			bookmarksEntry.getUserUuid());
 
-		BookmarksFolder existingFolder = fetchStagedModelByUuidAndGroupId(
-			bookmarksFolder.getUuid(), portletDataContext.getScopeGroupId());
+		BookmarksEntry existingBookmarksEntry =
+			fetchStagedModelByUuidAndGroupId(
+				bookmarksEntry.getUuid(), portletDataContext.getScopeGroupId());
 
-		if ((existingFolder == null) || !isStagedModelInTrash(existingFolder)) {
+		if ((existingBookmarksEntry == null) ||
+			!isStagedModelInTrash(existingBookmarksEntry)) {
+
 			return;
 		}
 
-		TrashHandler trashHandler = existingFolder.getTrashHandler();
+		TrashHandler trashHandler = existingBookmarksEntry.getTrashHandler();
 
 		try {
-			if (trashHandler.isRestorable(existingFolder.getFolderId())) {
+			if (trashHandler.isRestorable(
+					existingBookmarksEntry.getEntryId())) {
+
 				trashHandler.restoreTrashEntry(
-					userId, existingFolder.getFolderId());
+					userId, existingBookmarksEntry.getEntryId());
 			}
 		}
 		catch (PortalException pe) {
