+++ modules/apps/bookmarks/bookmarks-service/src/com/liferay/bookmarks/lar/BookmarksEntryStagedModelDataHandler.java
@@ -19,7 +19,7 @@ import com.liferay.bookmarks.model.BookmarksFolder;
 import com.liferay.bookmarks.model.BookmarksFolderConstants;
 import com.liferay.bookmarks.service.BookmarksEntryLocalServiceUtil;
 import com.liferay.exportimport.lar.BaseStagedModelDataHandler;
-import com.liferay.exportimport.stagedmodel.repository.StagedModelRepository;
+import com.liferay.exportimport.staged.model.repository.StagedModelRepository;
 import com.liferay.portal.kernel.util.MapUtil;
 import com.liferay.portal.kernel.xml.Element;
 import com.liferay.portal.service.ServiceContext;
